Feat: 3-06 회원가입 인코딩 없이

+++ src/test/java/com/ll/exam/sbb/QuestionRepositoryTests.java
@@ -19,7 +19,7 @@ public class QuestionRepositoryTests {
     private QuestionRepository questionRepository;
     @Autowired
     private QuestionService questionService;
-    private static int lastSampleDataId;
+    private static Long lastSampleDataId;
 
     @BeforeEach
     void beforeEach() {
@@ -37,7 +37,7 @@ public class QuestionRepositoryTests {
         }
     }
 
-    public static int createSampleData(QuestionRepository questionRepository) {
+    public static Long createSampleData(QuestionRepository questionRepository) {
         Question q1 = new Question();
         q1.setSubject("sbb가 무엇인가요?");
         q1.setContent("sbb에 대해서 알고 싶습니다.");
@@ -88,7 +88,7 @@ public class QuestionRepositoryTests {
     void 삭제() {
         assertThat(questionRepository.count()).isEqualTo(lastSampleDataId);
 
-        Question q = this.questionRepository.findById(1).get();
+        Question q = this.questionRepository.findById(1L).get();
         questionRepository.delete(q);
 
         assertThat(questionRepository.count()).isEqualTo(lastSampleDataId - 1);
@@ -96,11 +96,11 @@ public class QuestionRepositoryTests {
 
     @Test
     void 수정() {
-        Question q = this.questionRepository.findById(1).get();
+        Question q = this.questionRepository.findById(1L).get();
         q.setSubject("수정된 제목");
         questionRepository.save(q);
 
-        q = this.questionRepository.findById(1).get();
+        q = this.questionRepository.findById(1L).get();
 
         assertThat(q.getSubject()).isEqualTo("수정된 제목");
     }
